refactor path cleaning toggle into its own thing
